[FIX][DCE_ENC_TEST] Allocate buffer for Motion vector and SAD
enableAnalyticinfo is enabled by default for MPEG4/H263 encode and the required MVInfo buffer is not allocated.
This leads to memory corruption in MPEG4/H263 codec for encode with certain resolutions on QNX platform.
Buffer handling is added for MVInfo buffer when enableAnalyticinfo is enabled.
enableAnalyticinfo is disabled by default for MPEG4/H263 encode case.

For Internal reference, it fixes the  DRT OMAPS00300892 - 720P video encode is not functional with MPEG4/H263

Change-Id: I13d5d61faf8352d884070c3e1f17bb7005157334
Signed-off-by: Subash Lakkimsetti <x0091084@ti.com>