[LIBDCE]: Changes to make Encoder and Decoder co-exist on Android
In the current design omx video decoder is sending ptr for
inbufs and dma_buf_fd for the outbufs to the dce client.
In order to co-exist with encoder the inbufs/outbufs interpretation
is unified between encoder and decoder. this is achieved by
passing the memheader for both inbufs and outbufs for
both encoder and decoders.
This patch modified the process function to handle memheader
for all Android scenarios (inbufs/outbufs for encoder/decoder)
With this unified buf interpretation, the process function
simplified for Android.

Note# This patch has should be merged along with the omx patch:
http://review.omapzoom.org/#/c/34414/

Change-Id: Ia89ef9785ddf702bca986ce6b4a91983fb8bcc11
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>