[Bug] Fix possible client count overflow
The number of existing clients already running was being incremented
and then checked for maximum clients allowed.
However in the error scenario i.e. when maximum allowed clients are already
running and the new incoming client connection request was rejected the
client count was incremented unconditionally and not rolled back to
previous value thus resulting in possible overflow error.

This patch fixes this erroneous case.

Change-Id: I3e2b2bebfd025ba6a99b7ef838537d9716154ca4
Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>