[DCE_TEST] When outBufsInUse is set, use previous inputID
When handling interlace clip, the codec can hold the output buffer after
processing the top field frame. The codec will set the outBufsInUse flag.
When the application receives an indication that outBufsInUse flag is set,
the application should use the previous inputID as the inputID and not
sending a new output buffer (set to NULL on the output buffer) on the next
VIDDEC3_process call.

Change-Id: Id5e05984330e686ec5ea449be5d0aff8f8802d70
Signed-off-by: buddy.liong <buddy.budiono@ti.com>