Fix multi threaded logic for client indexing
Client count was being used for indexing into an empty slot in table
holding clients' engine handles.

In the case of multi-threaded scenario, there is a possibility of
out of order creation and deletions of clients in which case the client
indexing should search for an empty slot in the table rather than assuming
that the index defined by number of clients points to an empty slot in the
table.

This patch fixes this issue by decoupling client count from table indexing.
Now the table indexing is done by finding the empty slot identified by a '0'
in the entry. This also forces filling '0' during client deletion which is also
taken care of in the patch.

Change-Id: I3d4c8d744c3912a6fae5809f445039d39b762f91
Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>