[DCE_TEST] Fixing the decoded output resolution
Previously, when decoding for example H.264 1080p (1920x1080),
the output yuv comes out to be 1920x1088 because codec requires
the maxWidth and maxHeight on the VIDDEC3_Params to be aligned by 16 bytes.

Problem is in dce_test.c, the aligned is done early to the global variable
width and height which altered the value to be aligned by 16 bytes.
On the example above the height becomes 1088 (aligned by 16 bytes).
When the output buffer is allocated for TILER, the
outBufs->descs[0].tileMem.width = width that is aligned to 16 bytes.
outBufs->descs[0].tileMem.height = height that is aligned to 16 bytes.
or when the output buffer is allocated for Non-TILER, the
outBufs->descs[0].bytes = width * height that is aligned to 16 bytes.

With this fix, the output is now corrected to have the same resolution
as the input.
Eg. H.264 1080p (1920x1080) will be decoded to YUV NV12 of 1920x1080.
No longer YUV NV12 of 1920x1088.

Change-Id: Icdce8dfb99ffd2568ad318ff30305f83901a8243
Signed-off-by: Buddy Liong <a0270631@ti.com>