get_rproc_info: Introduce new API to query remoteproc
This is the initial implementation of an extensible API
that allows the user to query various remote core
parameters. Currently, there are three query parameters
that are supported:
  1. CPU Load
  2. Total configured heap size
  3. Available heap size

The intent of this API is to programmatically call the
API and feed it in various other tools that can be
developed on top of this (for example: soc performance
visualization etc.)

NOTE: This will require the implementation on ipumm-fw
and/or dspdce-fw.

Change-Id: I94c309221dbce24162a4f31166890d3acdb2d049
Signed-off-by: Karthik Ramanan <a0393906@ti.com>
Signed-off-by: Buddy Liong <buddy.liong@ti.com>