memplugin android: Fix h->ptr offset in memplugin_alloc
In memplugin_alloc(), the buffer pointer is meant to start at the end of
the MemHeader structure. h->ptr which holds this offset is calculated
incorrectly, but this has not caused any errors as the h->ptr is not
referenced directly. This offset is acquired through `H2P` macroin the
code. Nevertheless fixing the pointer arithmetic.

Change-Id: I075455860fbeb54001a9f2fde1de544c39fe3ed0
Signed-off-by: Gowtham Tammana <g-tammana@ti.com>