libdce[Android]: Fix NULL pointer dereference in dce_buf_unlock
In the case where dce_buf_lock fails due to dce_ipc_init failure,
then MmRpcHandle is not guaranteed to be initialized and could
result in a NULL pointer de-reference. In some use-cases where the
IPUMM takes a longer time to come up, this was resulting in a crash,
causing the Android media service to stop and restart several
times.

This patch fixes the NULL pointer de-reference by first checking if
the IPC is initialized in dce_buf_unlock before calling the IPC API.

This patch also has a minor cleanup of an un-used static global variable.

Change-Id: Idb74fb60c0b9696a0be87e4808b537ebfc84cd7e
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>