libdce [Android]: Increase array size for inbuf_offset and outbuf_offset
For Android only, in process(), it will use the inbuf_offset array
and outbuf_offset array when marshalling functions arguments to be
sent to remote core.

Before this patch, MAX_INPUT_BUF and MAX_OUTPUT_BUF are set to 2
meaning inbuf_offset and outbuf_offset are only an array of 2 instances.

With the changes in OMX to enable Error Concealment on MPEG2, which
requires an additional meta buffer to be passed in outBufs argument,
the outbuf_offset is no longer sufficient.

This commit changes the array number to 16 which is based on the
XDM_MAX_IO_BUFFERS value for XDM2_BufDesc.

For internal reference: CSR OMAPS00343473.

Change-Id: Ia490ef03017e4c4cdb82cda1a22cbc8b3f28567d
Signed-off-by: Buddy Liong <buddy.liong@ti.com>